refactor[src/constants]: Create a constants and enums directorie
