+++ src/frontend/src/components/promptComponent/index.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from "react";
 
-import { TypeModal } from "../../constants";
+import { TypeModal } from "../../constants/constants";
 import { PopUpContext } from "../../contexts/popUpContext";
 import { typesContext } from "../../contexts/typesContext";
 import { postValidatePrompt } from "../../controllers/API";
