+++ src/frontend/src/modals/formModal/chatMessage/codeBlock/index.tsx
@@ -2,7 +2,7 @@ import { IconCheck, IconClipboard, IconDownload } from "@tabler/icons-react";
 import { useState } from "react";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { oneDark } from "react-syntax-highlighter/dist/cjs/styles/prism";
-import { programmingLanguages } from "../../../../constants";
+import { programmingLanguages } from "../../../../constants/constants";
 
 interface Props {
   language: string;
